Move methods for transactions to a single class
