+++ src/newbank/server/Transactions.java
@@ -1,0 +1,4 @@
+package newbank.server;
+
+public class Transactions {
+}
